+++ django_production/settings.py
@@ -1,5 +1,6 @@
 import os
 import re
+from pathlib import Path
 
 import environ
 
@@ -7,8 +8,12 @@ env = environ.FileAwareEnv()
 
 DJANGO_ENV = env.str("DJANGO_ENV", "dev")
 
-# Allow running webserver from manage.py
-INSTALLED_APPS.append("django_webserver")
+INSTALLED_APPS.extend(
+    [
+        "django_webserver",  # Allow running webserver from manage.py
+        "whitenoise.runserver_nostatic",  # Use whitenoise with runserver
+    ]
+)
 
 # Insert whitenoise middleware
 try:
@@ -25,8 +30,19 @@ MIDDLEWARE.insert(0, "django_alive.middleware.healthcheck_bypass_host_check")
 X_FRAME_OPTIONS = "DENY"
 REFERRER_POLICY = "same-origin"
 
+# Staticfiles
+staticfiles_parent_dir = Path(__file__).resolve().parents[$settings_depth]
+STATICFILES_STORAGE = "whitenoise.storage.CompressedManifestStaticFilesStorage"
+STATICFILES_DIRS = [staticfiles_parent_dir / "static"]
+STATIC_ROOT = staticfiles_parent_dir / "static_collected"
+WHITENOISE_AUTOREFRESH = True
+WHITENOISE_IMMUTABLE_FILE_TEST = lambda path, url: re.match(  # noqa: E731
+    r"^.+\.[0-9a-f]{12}\..+$", url
+)
+
 if DJANGO_ENV == "production":
     DEBUG = env.bool("DEBUG", default=False)
+    WHITENOISE_AUTOREFRESH = False
     SECRET_KEY = env.str("SECRET_KEY")
     if "DATABASE_URL" in os.environ:
         DATABASES = {"default": env.db("DATABASE_URL")}
@@ -51,11 +67,4 @@ if DJANGO_ENV == "production":
     # Only set this to True if you are certain that all subdomains of your domain should be served exclusively via SSL.
     SECURE_HSTS_INCLUDE_SUBDOMAINS = env.bool(
         "SECURE_HSTS_INCLUDE_SUBDOMAINS", default=False
-    )
-
-    # Staticfiles
-    STATICFILES_STORAGE = "whitenoise.storage.CompressedManifestStaticFilesStorage"
-    # Match filename with 12 hex digits before the extension
-    WHITENOISE_IMMUTABLE_FILE_TEST = lambda path, url: re.match(  # noqa: E731
-        r"^.+\.[0-9a-f]{12}\..+$", url
     )